requirements.txt
semantic-kernel[azure]
fastapi
uvicorn
azure-monitor-opentelemetry-exporter
